third_party/WebKit/LayoutTests/editing/selection/readonly-disabled-text-selection.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
onload = function() {
    if (!window.eventSender || !eventSender.gestureLongPress) {
        debug("gestureLongPress not implemented by this platform.");
        debug("Manullay long press on every element in the page and check whether Text selection is happening or not");
        debug("If Text selection is not happening for readonly or disabled input/textarea, then it's a failure.");
        return;
    }

    doLongPressOnElement("normalText");

    doLongPressOnElement("readOnlyText");

    doLongPressOnElement("disabledText");

    doLongPressOnElement("readOnlyDisabledText");

    doLongPressOnElement("normalTextArea");

    doLongPressOnElement("readOnlyTextArea");
					
    doLongPressOnElement("disabledTextArea");

    doLongPressOnElement("readOnlyDisabledTextArea");
}

function doLongPressOnElement(elementId) {
    var element = document.getElementById(elementId);
    var bounds = element.getBoundingClientRect();
    var x = bounds.left + 5;
    var y = bounds.top + 5;
    eventSender.gestureLongPress(x, y);
    shouldBeEqualToString('window.getSelection().toString()', element.value);
}
</script>
<input id="normalText" type="text" value="NormalInput">
<input id="readOnlyText" type="text" value="ReadonlyInput" readonly>
<input id="disabledText" type="text" value="DisabledInput" disabled>
<input id="readOnlyDisabledText" size="20" type="text" value="ReadonlyDisabledInput"readonly disabled>
<textarea id="normalTextArea" cols="31">NormalTextarea</textarea>
<textarea id="readOnlyTextArea" cols="31" readonly>ReadonlyTextarea</textarea>
<textarea id="disabledTextArea" cols="31" disabled>DisabledTextarea</textarea>
<textarea id="readOnlyDisabledTextArea" cols="31" readonly disabled>ReadonlyDisabledTextarea</textarea>
